Extreme performance gain from multithreaded KeyboardListener + KeyboardInputBuffer: physics frametime: 9.29ms to 0.0459ms visuals frametime: 18.58ms to 15.24ms
Since the KeyboardInputBuffer uses raw queue, there will be a chance of the program failing. An implementation for Single Producer, Single Consumer (SPSC or SCSP) queue is needed to minimize this chance.

Also noticed the threads of StdTxtRenderer constantly exiting and creating new threads. Synchronize the threads to reduce threading overhead?